Move to item_events module
