+++ library/item/item_events/src/main/java/org/quiltmc/qsl/item/events/api/QuiltItemUsageContextExtensions.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.item.extensions.api;
+package org.quiltmc.qsl.item.events.api;
 
 import org.jetbrains.annotations.NotNull;
