+++ library/item/item_events/src/main/resources/assets/quilt_item_events/icon.png
