+++ library/item/item_events/src/main/resources/quilt_item_events.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "org.quiltmc.qsl.item.extensions.mixin",
+  "package": "org.quiltmc.qsl.item.events.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
     "ItemStackMixin",
