+++ library/item/item_events/src/testmod/resources/quilt.mod.json
@@ -2,25 +2,25 @@
   "schema_version": 1,
   "quilt_loader": {
     "group": "org.quiltmc.qsl.item",
-    "id": "quilt_item_extensions_testmod",
+    "id": "quilt_item_events_testmod",
     "version": "1.0.0",
     "metadata": {
-      "name": "Quilt Item Extensions Test Mod",
+      "name": "Quilt Item Events Test Mod",
       "license": "Apache-2.0"
     },
     "intermediate_mappings": "net.fabricmc:intermediary",
     "load_type": "always",
     "entrypoints": {
       "init": [
-        "org.quiltmc.qsl.item.extensions.test.Initializer::INSTANCE"
+        "org.quiltmc.qsl.item.events.test.Initializer::INSTANCE"
       ],
       "events": [
-        "org.quiltmc.qsl.item.extensions.test.Initializer::INSTANCE"
+        "org.quiltmc.qsl.item.events.test.Initializer::INSTANCE"
       ]
     },
     "depends": [
       "quilt_loader",
-      "quilt_item_extensions",
+      "quilt_item_events",
       "quilt_resource_loader"
     ]
   }
