--- library/item/item_extensions/build.gradle
@@ -1,21 +1,0 @@
-plugins {
-	id("qsl.module")
-}
-
-qslModule {
-	name = "Quilt Item Extensions API"
-	moduleName = "item_extensions"
-	id = "quilt_item_extensions"
-	description = "Extensions for creating and working with items."
-	library = "item"
-	moduleDependencies {
-		core {
-			api("qsl_base")
-			testmodOnly("resource_loader")
-		}
-	}
-
-	injectedInterface("net/minecraft/class_1838") {
-		values = ["org/quiltmc/qsl/item/extensions/api/QuiltItemUsageContextExtensions"]
-	}
-}
